fix type of Iter
Former-commit-id: 4e461fd18cb0bf9044c6b5aee3311867bbd746d4 [formerly b94a8a0eb038ff13bcf5e1be96dca3ce77c8f229]
Former-commit-id: 4637b53946846af9a59c4995456a3298ea343d82

+++ Examples/CapitalConstraints/GlobalSolution.m
@@ -36,14 +36,14 @@ function [ V, C, CB, W, kv, alpha, beta, nu, theta ] = GlobalSolution
     CB = ( exp( ( 1 + nu ) * ( ag + alpha * kg ) ) * ( 1 - alpha ) ^ ( 1 - alpha ) ) .^ ( 1 / ( 1 + nu ) );
 
     fprintf( '\n' );
-    Iter = 0;
+    Iter = int32( 0 );
     while true
         thetac = min( theta, Iter / 100 );
         [ Vnew, Cnew, CBnew ] = IterateValueFunction( V, C, CB, W, kv, av, alpha, beta, nu, thetac );
         e1 = max( max( abs( V - Vnew ) ) );
         e2 = max( max( abs( C - Cnew ) ) );
         e3 = max( max( abs( CB - CBnew ) ) );
-        Iter = Iter + 1;
+        Iter = Iter + int32( 1 );
         fprintf( '%d %.15g %.15g %.15g %.15g\n', Iter, thetac, e1, e2, e3 );
         V = Vnew;
         C = Cnew;
